% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{add_snp_to_annotation_track}
\alias{add_snp_to_annotation_track}
\title{add_snp_to_annotation_track
Add SNP information to an annotation track.
NOTE: SNPS could be made as a separate track. However, additional subplots 
compress
the vertical space, and when there is insufficient room the hoverinfo goes 
away.}
\usage{
add_snp_to_annotation_track(ann_track, snp_info, group_col = "CONTEXT")
}
\arguments{
\item{ann_track}{plotly: annotation track}

\item{snp_info}{GRanges: contains SNP information. Currently supports GRanges
made from GWAS catalog}

\item{group_col}{character: the column in snp_info for grouping traces. 
Default is "CONTEXT" which corresponds to the type of mutation in the GWAS 
catalog, e.g. 'missense'}
}
\value{
plotly_object
}
\description{
add_snp_to_annotation_track
Add SNP information to an annotation track.
NOTE: SNPS could be made as a separate track. However, additional subplots 
compress
the vertical space, and when there is insufficient room the hoverinfo goes 
away.
}
\author{
Justin Finkle
}
