% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{add_tx_shapes}
\alias{add_tx_shapes}
\title{add_tx_shapes
Add annotation shapes to a specific axis}
\usage{
add_tx_shapes(plotly_obj, tx_info, ann_ax, target_range)
}
\arguments{
\item{plotly_obj}{plotly: object to be modified}

\item{tx_info}{GRanges: contains annotation information. Must have metadata column with name 'feature' to make shapes}

\item{ann_ax}{character: axis to add shapes to, of the form 'yaxis[2-9]'}

\item{target_range}{GRanges: range to display. Used to draw shapes properly}
}
\description{
add_tx_shapes
Add annotation shapes to a specific axis
}
\author{
Justin Finkle
}
