% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{add_tx_stepping}
\alias{add_tx_stepping}
\title{add_tx_stepping
Set the level that that transcripts will be displayed on}
\usage{
add_tx_stepping(tx_gr, stacking = c("dense", "squish"))
}
\arguments{
\item{tx_gr}{GRanges: contains transcript information}

\item{stacking}{str: dense collapses transcripts, squish expands them}
}
\value{

}
\description{
add_tx_stepping
Set the level that that transcripts will be displayed on
}
\author{
Justin Finkle
}
