% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_plus_heatmap_shiny.R
\name{aggregate_to_heatmap_shiny}
\alias{aggregate_to_heatmap_shiny}
\title{aggregate_to_heatmap_shiny}
\usage{
aggregate_to_heatmap_shiny(aggregate_plot, heatmaps,
  title = "Aggregate Plot Linked to Heatmap", options = list(height = 1400))
}
\description{
aggregate_to_heatmap_shiny
}
