% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_plus_heatmap_shiny.R
\name{aggregate_to_heatmap_to_browserly_shiny}
\alias{aggregate_to_heatmap_to_browserly_shiny}
\title{heatmap_to_browserly_shiny}
\usage{
aggregate_to_heatmap_to_browserly_shiny(aggregate_plot, heatmaps,
  track_function, link_functions, title = "Heatmap linked to Genome Tracks",
  options = list(height = 1400))
}
\description{
heatmap_to_browserly_shiny
}
