% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{browserly_annotation_track}
\alias{browserly_annotation_track}
\title{browserly_annotation_track
Make the invisible points for an annotation track.}
\usage{
browserly_annotation_track(tx_info, track_name = "Annotation", opacity = 0,
  markercolor = "grey")
}
\arguments{
\item{tx_info}{GRanges: mcols describe the annotation features}

\item{track_name}{character: name for the track. May be used for referencing 
axes and adjusting domains}

\item{opacity}{numeric: value between 0 and 1. Sets the opacity of the points 
used for hoverinfo. Should almost always be 0.}

\item{markercolor}{character: color to use for markers. Default is 'grey'. If 
set to NULL, different colors will be used for each class of anntoation 
feature, e.g. intron, cds, utr3}
}
\value{
plotly_object
}
\description{
Currently this is only a helper function, but it could be made to stand alone
}
\author{
Justin Finkle
}
