% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_potential.R
\name{compute_regulatory_potential}
\alias{compute_regulatory_potential}
\title{compute_regulatory_potential}
\usage{
compute_regulatory_potential(peaks, tss, max_dist = 1e+05)
}
\arguments{
\item{peaks}{GRanges}

\item{tss}{GRanges or GRangesList}

\item{max_dist}{maximum distance between peak and tss for inclustion in regulatory 
potential score}
}
\value{
vector of length tss, with regulatory potential score for each tss
}
\description{
compute_regulatory_potential
}
\author{
Alicia Schep
}
