% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{coverage_heatmap}
\alias{coverage_heatmap}
\alias{coverage_heatmap,matrix-method}
\alias{coverage_heatmap,list-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,matrix-method}
\alias{add_coverage_heatmap}
\title{coverage_heatmap}
\usage{
coverage_heatmap(data, ...)

add_coverage_heatmap(p, data, ...)
}
\arguments{
\item{data}{single coverage matrix or list of coverage matrices}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{clust_dist}{distance function to use for clustering}

\item{signal}{signal along row}

\item{plot_signal}{add an annotation heatmap showing the average signal? 
default is TRUE}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean, default is TRUE}

\item{scale_method}{how to scale matrix before displaying in heatmap}

\item{pct}{percentile to use if scale_method is "PercentileMax"}

\item{scale_factor}{scale_factor to use if scale_method is "scalar"}

\item{show_xlabels}{show xlabels?  default is TRUE}

\item{start}{label for start of x range}

\item{end}{label for end of x range}

\item{xlab}{x axis label}

\item{font}{list of font attributes}
}
\value{
iheatmap object
}
\description{
coverage_heatmap
}
\author{
Alicia Schep
}
