% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{.fillwithNA}
\alias{.fillwithNA}
\title{fill the gaps for segment-level SCNA data with NA}
\usage{
.fillwithNA(segDf, totalChrRangeObj)
}
\arguments{
\item{segDf}{segment-level SCNA data frame. This is usually the same input as GISTIC2. Make sure there are these columns "sample", "chromosome", "start", "end", "log2".}

\item{totalChrRangeObj}{a GenomicRanges object, specifying the lengths of all chromosomes}
}
\value{
a data frame with gap-filled segment-level SCNA data
}
\description{
fill the gaps for single sample's segment-level SCNA data with NA
}
