% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{.seg2Ranges}
\alias{.seg2Ranges}
\title{convert a data frame of chromosome length to GRanges Obj}
\usage{
.seg2Ranges(totalChrLengthDf)
}
\arguments{
\item{totalChrLengthDf}{}
}
\value{
a GenomicRanges Obj
}
\description{
convert a data frame of chromosome length to GRanges Obj
}
