% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{.selectRange.2}
\alias{.selectRange.2}
\title{Subset genomic ranges from multiple samples' segment-level SCNA data}
\usage{
.selectRange.2(
  segDf,
  chr = NULL,
  start = NULL,
  end = NULL,
  genomeVersion = "hg38"
)
}
\arguments{
\item{segDf}{segment-level SCNA data frame containing one or multiple samples. This is usually the same input as GISTIC2.
Make sure there are these columns: "sample", "chromosome", "start", "end", "log2".}

\item{chr}{a character vector specifying which chromosome(s) to select}

\item{start}{a integer vector specifying the start positions of each chromosome.
If NULL, the position will be the begining of the chromosome(s). Default is NULL.}

\item{end}{a integer vector specifying the end positions of each chromosome.
If NULL, the position will be the end of the chromosome(s). Default is NULL.}

\item{genomeVersion}{which genomeVersion are you working on? Can be either 'hg38' and 'hg19'. Default is 'hg38'.}
}
\value{
a data frame representing the subsetted segment-level SCNA data from the input.
}
\description{
Subset genomic ranges from multiple samples' segment-level SCNA data
}
