% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_range.R
\name{fetch_view_range}
\alias{fetch_view_range}
\title{fetch_view_range}
\usage{
fetch_view_range(db_object, symbol, keytype = c("SYMBOL", "TXNAME"),
  relative = c("TSS", "TTS", "full"), up = 1000, down = 1000)
}
\arguments{
\item{db_object}{OrganismDb or TxDb object}

\item{symbol}{name of gene}

\item{keytype}{type of key}

\item{relative}{get range relative to TSS, TTS, or full gene}

\item{up}{basepairs upstream to include}

\item{down}{basepairs downstream to include}
}
\value{
GenomicRanges
}
\description{
Function to get a range around a gene, using an OrganismDb or TxDb object
}
\examples{

library(Homo.sapiens)
fetch_view_range(Homo.sapiens,"GLI2")
}
\author{
Alicia Schep and Justin Finkle
}
