% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{GenomicWidgets-shiny}
\alias{GenomicWidgets-shiny}
\alias{GenomicWidgetsOutput}
\alias{renderGenomicWidgets}
\title{Shiny bindings for GenomicWidgets}
\usage{
GenomicWidgetsOutput(outputId, width = "100\%", height = "400px")

renderGenomicWidgets(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates an Iheatmap object}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This 
is useful if you want to save an expression in a variable.}
}
\value{
analagous to \code{\link[shiny]{renderPlot}} and 
  \code{\link[shiny]{plotOutput}} from shiny package, renderGenomicWidgets 
  is used in server side code to render the plot, and GenomicWidgetsOutput
  is used within the ui code.
}
\description{
Output and render functions for using GenomicWidgets within Shiny
}
