% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicWidgets.R
\docType{package}
\name{GenomicWidgets}
\alias{GenomicWidgets}
\alias{GenomicWidgets-package}
\title{GenomicWidgets}
\description{
Interactive Visualizations for Functional Genomics
}
\details{
See the vignette for detailed instructions for how to use the package.

GenomicWidgets uses the plotly javascript library (\url{plot.ly}) for making the 
interactive figures and htmlwidgets (http://www.htmlwidgets.org/) for 
rendering them in R.
}
\author{
Alicia Schep, Justin Finkle, and Sarah Kummerfeld
}
