% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{get_annotation_axis}
\alias{get_annotation_axis}
\title{get_annotation_axis}
\usage{
get_annotation_axis(ax_info, annotation_str = "Annotation")
}
\arguments{
\item{ax_info}{data.frame like: contains info for different subplot axes in 
plotly object}

\item{annotation_str}{str: the string used to find the annotation axis}
}
\value{
str: the annotation axis reference in the form "yaxis[2-9]"
}
\description{
get_annotation_axis
}
\author{
Justin Finkle
}
