% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{get_centered_gene_info}
\alias{get_centered_gene_info}
\title{get_centered_gene_info
Center coverage and annotation around TSS
NOTES: 1. Can be generalized to pass target ranges instead of gene list
       2. Can be generalized to center around any feature
       3. Probably should only pass annotations relevant to the genes}
\usage{
get_centered_gene_info(gene, extension = 50000, cvg_files, sample_names,
  scaling_factor, tx_data, symbol_table)
}
\arguments{
\item{gene}{character: gene symbol to be centered}

\item{extension}{integer: upstream (and downstream) region around center to 
be viewed}

\item{cvg_files}{characters: paths to coverage files of interest}

\item{sample_names}{characters: names for the cvg_files to be displayed}

\item{scaling_factor}{numeric: scaling factor for coverage, e.g. library size}

\item{tx_data}{list: annotation information}

\item{symbol_table}{data.frame: annotation information linking gene symbols 
with txdb information}
}
\description{
get_centered_gene_info
Center coverage and annotation around TSS
NOTES: 1. Can be generalized to pass target ranges instead of gene list
       2. Can be generalized to center around any feature
       3. Probably should only pass annotations relevant to the genes
}
\author{
Justin Finkle
}
