% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_expression_set}
\alias{get_expression_set}
\title{Get expression set object. Either load from rds or supply expression plot 
project that can be retrieved}
\usage{
get_expression_set(filename, stat = NULL)
}
\arguments{
\item{filename}{string: preferred method -- provide rds for ExpressionSet object. 
Can also provide the project name to retrieve from ExpressionPlot}

\item{stat}{}
}
\description{
Get expression set object. Either load from rds or supply expression plot 
project that can be retrieved
}
