% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{get_snps_in_range}
\alias{get_snps_in_range}
\title{get_snps_in_range
Get the SNPs that fall in a particluar range}
\usage{
get_snps_in_range(snp_gr = GRanges(), target_range = GRanges())
}
\arguments{
\item{snp_gr}{GRanges: contains SNP information, e.g. from GWAS catalog. See 
load_snp_data for more info}

\item{target_range}{GRanges: the range in which to look for SNPs}
}
\description{
get_snps_in_range
Get the SNPs that fall in a particluar range
}
\author{
Justin Finkle
}
