% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plus_tracks_shiny.R
\name{heatmap_to_tracks_shiny}
\alias{heatmap_to_tracks_shiny}
\title{heatmap_to_tracks_shiny}
\usage{
heatmap_to_tracks_shiny(heatmap, track_function, link,
  title = "Heatmap linked to Genome Tracks", options = list(height = 1400))
}
\arguments{
\item{heatmap}{IHeatmap object}

\item{track_function}{track function generator}

\item{link}{link function, linking rows of heatmap to input to track function
generator}

\item{title}{Title of shiny app}

\item{options}{to pass to shiny}
}
\value{
Shiny application
}
\description{
Function for making shiny app linking a heatmap to genome tracks
}
\author{
Alicia Schep and Justin Finkle
}
