% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_iheatmap.R
\docType{methods}
\name{iheatmap-SummarizedExperiment}
\alias{iheatmap-SummarizedExperiment}
\alias{add_iheatmap,IheatmapHorizontal,SummarizedExperiment-method}
\alias{add_iheatmap,IheatmapVertical,SummarizedExperiment-method}
\title{Summarized Experiment iheatmap and iheatmap methods}
\usage{
\S4method{iheatmap}{SummarizedExperiment}(data, assay = assayNames(data)[[1]],
  x = se_default_x(data, assay), y = se_default_y(data, assay),
  cluster_rows = c("hclust", "kmeans", "none"), cluster_cols = c("hclust",
  "kmeans", "none"), scale = "rows", ...)

\S4method{add_iheatmap}{IheatmapHorizontal,SummarizedExperiment}(p, data,
  assay = assayNames(data)[[1]], x = se_default_x(data),
  cluster_cols = c("hclust", "kmeans", "none"), scale = "rows", ...)

\S4method{add_iheatmap}{IheatmapVertical,SummarizedExperiment}(p, data,
  assay = assayNames(data)[[1]], x = se_default_x(data),
  cluster_rows = c("hclust", "kmeans", "none"), scale = "rows", ...)
}
\arguments{
\item{data}{SummarizedExperiment}

\item{assay}{name of assay in data to plot}

\item{x}{name of rows of data, by default rownames}

\item{y}{name of columns of data, by default colnames}

\item{cluster_rows}{how to cluster rows, default is hclust}

\item{cluster_cols}{how to cluster columns, default is hclust}

\item{...}{additional arguments to \code{\link[iheatmapr]{iheatmap}}}

\item{row_annotation}{annotation data frame for rows}

\item{col_annotation}{annotation data frame for columns}
}
\description{
Summarized Experiment iheatmap and iheatmap methods
}
