% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTrackWidget.R
\name{knit_print.GenomeTrackWidget}
\alias{knit_print.GenomeTrackWidget}
\title{knit_print.GenomeTrackWidget}
\usage{
\method{knit_print}{GenomeTrackWidget}(x, options)
}
\arguments{
\item{x}{GenomeTrackWidget object}

\item{options}{knitr options}
}
\description{
knit_print.GenomeTrackWidget
}
\keyword{internal}
