% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTrackWidget.R
\name{knit_print.GenomicWidget}
\alias{knit_print.GenomicWidget}
\title{knit_print.GenomicWidget}
\usage{
\method{knit_print}{GenomicWidget}(x, options)
}
\arguments{
\item{x}{GenomicWidget object}

\item{options}{knitr options}
}
\description{
knit_print.GenomicWidget
}
\keyword{internal}
