% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTrackWidget.R
\name{knit_print.LocusSummaryList}
\alias{knit_print.LocusSummaryList}
\title{knit_print.LocusSummaryList}
\usage{
\method{knit_print}{LocusSummaryList}(x, options)
}
\arguments{
\item{x}{LocusSummaryList object}

\item{options}{knitr options}
}
\description{
knit_print.LocusSummaryList
}
\keyword{internal}
