% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTrackWidget.R
\name{knit_print.LocusViewList}
\alias{knit_print.LocusViewList}
\title{knit_print.LocusViewList}
\usage{
\method{knit_print}{LocusViewList}(x, options)
}
\arguments{
\item{x}{LocusViewList object}

\item{options}{knitr options}
}
\description{
knit_print.LocusViewList
}
\keyword{internal}
