% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTrackWidget.R
\name{knit_print_GenomeWidgets}
\alias{knit_print_GenomeWidgets}
\alias{knit_print.GenomeTrackWidget}
\alias{knit_print_GenomeWidgets#'}
\alias{knit_print.LocusViewList}
\alias{knit_print_GenomeWidgets#'}
\alias{knit_print.LocusSummaryList}
\title{knit_print for GenomicWidgets objects}
\usage{
\method{knit_print}{GenomeTrackWidget}(x, options)

\method{knit_print}{LocusViewList}(x, options)

\method{knit_print}{LocusSummaryList}(x, options)
}
\arguments{
\item{x}{GenomeTrackWidget object}

\item{options}{knitr options}
}
\value{
htmlwidgets object
}
\description{
knit_print method for objects that generate GenomicWidgets
}
\keyword{internal}
