% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{load_tx_data}
\alias{load_tx_data}
\title{Load Transcript Data}
\usage{
load_tx_data(db_object)
}
\arguments{
\item{db_object}{database object: currently accepts TxDB and OrganismDB}
}
\description{
Load Transcript Data
}
