% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{log10rowMeans}
\alias{log10rowMeans}
\title{log10rowMeans}
\usage{
log10rowMeans(x, na.rm = TRUE, pseudo = 1)
}
\arguments{
\item{x}{matrix}

\item{na.rm}{remove NAs?}

\item{pseudo}{pseudocount to add}
}
\value{
vector of values
}
\description{
Convenience function to take log10 after taking the mean. A pseudocount is 
added before taking the log to avoid problems with zero.
}
\examples{

mat = matrix(rpois(120,10), nrow = 10)
log10rowMeans(mat)

}
