% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{make_arrows}
\alias{make_arrows}
\title{make_arrows
Draw arrows for introns}
\usage{
make_arrows(df, yref, arrowlen = 500, arrowheight = 0.15, arrowgap = 1500)
}
\arguments{
\item{df}{data.frame: data used to draw the shapes. Required columns include 
"start", "end", "strand", "midpoint", and "stepping"}

\item{yref}{character: axis on which to draw the rectangles, in the form of 
"yaxis[2-9]" or "y[2-9]"}

\item{arrowlen}{numeric: how long the arrow should be, in x coordinates}

\item{arrowheight}{numeric: how tall the arrow should be, in yref coordinates}

\item{arrowgap}{numeric: gap between arrows on long introns}
}
\value{
list of arrow shapes
}
\description{
make_arrows
Draw arrows for introns
}
\author{
Justin Finkle
}
\keyword{interanl}
