% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_matrix.R
\name{make_coverage_tracks}
\alias{make_coverage_tracks}
\title{make_coverage_tracks}
\usage{
make_coverage_tracks(inputs, target_range, sample_names = names(inputs),
  binsize = 100, method = c("a", "j"), format = c("auto", "bigwig",
  "bam"), bin = TRUE, scaling_factors = rep(1, length(inputs)), up = 0,
  down = 0, stranded = FALSE)
}
\arguments{
\item{inputs}{filenames of bigwig, bam, or RData file}

\item{binsize}{binsize to bin coverage}

\item{format}{format of files, default is auto}

\item{up}{basepairs upstream of center to use}

\item{down}{basepairs downstream of center to use}

\item{ranges}{ranges for which to compute coverage within}
}
\description{
make_coverage_tracks
}
\details{
up and down are 0 by default -- if not specified, actual range is used.  All ranges
must be of equal width.  If up and/or down are provided, then the center of the range and up 
basepairs upstream and down basepairs downstream are used.
}
\author{
Alicia Schep
}
