% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{make_plotly_color}
\alias{make_plotly_color}
\title{make_plotly_color
Convert R colors to plotly compatible rgb values}
\usage{
make_plotly_color(color_str)
}
\arguments{
\item{color_str}{}
}
\description{
make_plotly_color
Convert R colors to plotly compatible rgb values
}
\author{
Justin Finkle
}
