% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{make_rect}
\alias{make_rect}
\title{make_rect
Make the rectangle objects. These are used to display annotation features such
as cds, and utr.}
\usage{
make_rect(df, height, yref, fillcolor = "lightslateblue")
}
\arguments{
\item{df}{data.frame: data used to draw the shapes. Required columns include 
"start", "end", and "stepping"}

\item{height}{numeric: the height of the rectangles}

\item{yref}{character: axis on which to draw the rectangles, in the form of 
"yaxis[2-9]" or "y[2-9]"}

\item{fillcolor}{character: color of the rectangle}
}
\value{
list of rectangle shapes
}
\description{
make_rect
Make the rectangle objects. These are used to display annotation features such
as cds, and utr.
}
\author{
Justin Finkle
}
\keyword{internal}
