% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_summary.R
\docType{methods}
\name{make_summary_plotter,SummarizedExperiment-method}
\alias{make_summary_plotter,SummarizedExperiment-method}
\title{make_summary_plotter}
\usage{
\S4method{make_summary_plotter}{SummarizedExperiment}(object,
  assay_name = assayNames(object)[1], ..., groups = NULL,
  showlegend = !is.null(colors), colors = NULL, boxpoints = c("all",
  "Outliers", "false"), pointpos = 0, ytitle = "Expression")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{assay_name}{name of assay to use}

\item{...}{additional arguments}

\item{groups}{either vector of group assignments of name of column in object 
colData that corresponds to vector of group assignments}

\item{showlegend}{show the legend?}

\item{colors}{colors to use}

\item{boxpoints}{plot individual points?}

\item{pointpos}{relative position of points to boxes}

\item{ytitle}{name for yaxis}
}
\description{
make_summary_plotter
}
