% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_summary.R
\docType{methods}
\name{make_summary_plotter}
\alias{make_summary_plotter}
\alias{make_summary_plotter,SummarizedExperiment-method}
\title{make_summary_plotter}
\usage{
\S4method{make_summary_plotter}{SummarizedExperiment}(object,
  assay_name = assayNames(object)[1], ..., groups = NULL,
  showlegend = !is.null(colors), colors = NULL, boxpoints = c("all",
  "Outliers", "false"), pointpos = 0, ytitle = "Expression")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{assay_name}{name of assay to use}

\item{...}{additional arguments}

\item{groups}{either vector of group assignments of name of column in object 
colData that corresponds to vector of group assignments}

\item{showlegend}{show the legend?}

\item{colors}{colors to use}

\item{boxpoints}{plot individual points?}

\item{pointpos}{relative position of points to boxes}

\item{ytitle}{name for yaxis}
}
\description{
Makes a summary plot for a genomic region. For use with make_track_plotter
and make_track_plus_summary_plotter.
}
\examples{

## we'll read in some RNA counts
data(rpkm_chr21)

## Make summary plotter

summary_plotter <- make_summary_plotter(rpkm_chr21,
  groups = "GROUP") 
  
if (interactive()){
  summary_plotter(rownames(rpkm_chr21)[1:3])
}   

}
\author{
Alicia Schep and Justin Finkle
}
