% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus.R
\docType{methods}
\name{make_track_plotter,character-method}
\alias{make_track_plotter,character-method}
\title{make_track_plotter}
\usage{
\S4method{make_track_plotter}{character}(object, annotation = NULL, ...,
  track_names = ifelse(!is.null(names(object)), names(object),
  basename(object)), groups = NULL, share_y = FALSE, fill = c("tozeroy",
  "none"), relative = FALSE, showlegend = TRUE, colors = NULL,
  mode = "lines", annotation_position = c("bottom", "top"),
  annotation_size = 0.25)
}
\arguments{
\item{object}{vector of bam or bigwig file names}

\item{annotation}{TxDb or OrganismDb object}

\item{...}{additional arguments}

\item{track_names}{names to associate with each file}

\item{groups}{vector of group assignments.  traces will be grouped onto subplots
based on group assignments (if only showing 1 region)}

\item{share_y}{share the y axis?}

\item{fill}{fillmode for line plot}

\item{relative}{ignore for now}

\item{showlegend}{show the legend?}

\item{colors}{colors for each bam file}

\item{mode}{mode for plot}

\item{annotation_position}{plot annotations on bottom or on top of signal traces}

\item{annotation_size}{relative size of annotation plot}
}
\description{
Function to generate a function that takes in a range and plots coverage track
}
