% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus.R
\docType{methods}
\name{make_track_plotter}
\alias{make_track_plotter}
\alias{make_track_plotter,character-method}
\title{make_track_plotter}
\usage{
\S4method{make_track_plotter}{character}(object, annotation = NULL, ...,
  track_names = ifelse(!is.null(names(object)), names(object),
  basename(object)), groups = NULL, share_y = FALSE, fill = c("tozeroy",
  "none"), relative = FALSE, showlegend = TRUE, colors = NULL,
  mode = "lines", annotation_position = c("bottom", "top"),
  annotation_size = 0.25)
}
\arguments{
\item{object}{vector of bam or bigwig file names}

\item{annotation}{TxDb or OrganismDb object}

\item{...}{additional arguments}

\item{track_names}{names to associate with each file}

\item{groups}{vector of group assignments.  traces will be grouped onto subplots
based on group assignments (if only showing 1 region)}

\item{share_y}{share the y axis?}

\item{fill}{fillmode for line plot}

\item{relative}{ignore for now}

\item{showlegend}{show the legend?}

\item{colors}{colors for each bam file}

\item{mode}{mode for plot}

\item{annotation_position}{plot annotations on bottom or on top of signal traces}

\item{annotation_size}{relative size of annotation plot}
}
\value{
function to make interactive track plots
}
\description{
Function to generate a function that takes in a range and plots coverage track
}
\examples{

library(GenomicRanges)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)

## First we'll read in some sample data
genomation_dir <- system.file("extdata", package = "genomationData")
samp.file <- file.path(genomation_dir,'SamplesInfo.txt')
samp.info <- read.table(samp.file, header=TRUE, sep="\\t", 
                        stringsAsFactors = FALSE)
samp.info$fileName <- file.path(genomation_dir, samp.info$fileName)
ctcf.peaks = genomation::readBroadPeak(system.file("extdata",
                         "wgEncodeBroadHistoneH1hescCtcfStdPk.broadPeak.gz",
                          package = "genomationData"))
ctcf.peaks = ctcf.peaks[seqnames(ctcf.peaks) == "chr21"]

## resize peaks to size 1000
ctcf.peaks = resize(ctcf.peaks, width = 10000, fix = "center")

## Make track plotter

track_plotter <- make_track_plotter(samp.info$fileName[1:3], 
  annotation = TxDb.Hsapiens.UCSC.hg19.knownGene, 
  track_names = samp.info$sampleName[1:3] , 
  share_y = TRUE)
if (interactive()){
  track_plotter(ctcf.peaks[1])
  track_plotter(ctcf.peaks[1:3])
}   

}
\author{
Alicia Schep and Justin Finkle
}
