% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus.R
\name{make_track_plus_summary_plotter}
\alias{make_track_plus_summary_plotter}
\title{Combine genome track view with locus summary}
\usage{
make_track_plus_summary_plotter(track_function, summary_function, windows,
  row_names, summary_width = 0.25)
}
\arguments{
\item{track_function}{function to make tracks, as created by
\code{\link{make_track_plotter}}}

\item{summary_function}{function to make locus summaries, as created by
\code{\link{make_summary_plotter}}}

\item{windows}{GenomicRanges of windows that can serve as inputs to 
track_function}

\item{row_names}{vector of row names that can serve as inputs to 
summary_function}

\item{summary_width}{width of summary plot in resulting visualization}
}
\value{
a function which takes in names and generates a plot with track views
and a summary plot per locus
}
\description{
Combine genome track view with locus summary
}
\examples{

library(GenomicRanges)
## First we'll read in some sample chip-seq data
genomation_dir <- system.file("extdata", package = "genomationData")
samp.file <- file.path(genomation_dir,'SamplesInfo.txt')
samp.info <- read.table(samp.file, header=TRUE, sep='\\t', 
                        stringsAsFactors = FALSE)
samp.info$fileName <- file.path(genomation_dir, samp.info$fileName)

## we'll also read in some RNA counts
data(rpkm_chr21)

## From the ranges of the rpkm object, we'll pull out the tss
tss <- promoters(rowRanges(rpkm_chr21), up = 1, down = 1)

## Make track plotter

track_plotter <- make_track_plotter(samp.info$fileName[1:3], 
  annotation = TxDb.Hsapiens.UCSC.hg19.knownGene, 
  track_names = samp.info$sampleName[1:3] , 
  share_y = TRUE)
  
## Make summary plotter

summary_plotter <- make_summary_plotter(rpkm_chr21,
  groups = "GROUP") 
  
## Make combined plotter
  
combined_plotter <- make_track_plus_summary_plotter(track_plotter, 
  summary_plotter,
  resize(tss, width = 5000, fix = "center"),
  rownames(rpkm_chr21))
  
if (interactive()){
  combined_plotter(rownames(rpkm_chr21)[1:3])
}   

}
\author{
Alicia Schep and Justin Finkle
}
