% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_range.R
\name{make_view_range}
\alias{make_view_range}
\title{make_view_range}
\usage{
make_view_range(chrom, start, end, strand = c("*", "+", "-"),
  name = character(0), relative = FALSE)
}
\description{
Makes a "ViewRange" object
}
