% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_locus.R
\docType{methods}
\name{min,AnnotationPlot-method}
\alias{min,AnnotationPlot-method}
\alias{max,AnnotationPlot-method}
\alias{min,SignalPlot-method}
\alias{max,SignalPlot-method}
\alias{min,LocusView-method}
\alias{max,LocusView-method}
\alias{min,LocusViewList-method}
\alias{max,LocusViewList-method}
\title{max, min for GenomicWidgets objects}
\usage{
\S4method{min}{AnnotationPlot}(x, ..., na.rm = FALSE)
}
\arguments{
\item{...}{additional arguments}

\item{na.rm}{remove na}
}
\value{
numeric
}
\description{
max, min for GenomicWidgets objects
}
\keyword{internal}
