% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_matrix.R
\docType{methods}
\name{normalize_coverage_matrix,list-method}
\alias{normalize_coverage_matrix,list-method}
\title{normalize_coverage_matrix}
\usage{
\S4method{normalize_coverage_matrix}{list}(mats, method = c("localRms",
  "localMean", "localNonZeroMean", "PercentileMax", "scalar", "none"),
  pct = 0.95, scalar = NULL)
}
\arguments{
\item{method}{normalization method option, see Details}

\item{pct}{Percentile, only used if PercentileMax is method}

\item{scalar}{vector of scalars used for normalizing each mat, only 
used if scalar is method}
}
\description{
normalize_coverage_matrix
}
\details{
Normalization choices are "localRms", "localMean", 
"localNonZeroMean", "PercentileMax", "scalar", and "none".
}
\author{
Alicia Schep
}
