% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_interface.R
\docType{methods}
\name{plot_tracks,GenomicRanges-method}
\alias{plot_tracks,GenomicRanges-method}
\alias{plot_tracks}
\alias{plot_tracks,character-method}
\alias{plot_tracks,GenomicRanges-method}
\title{plot_tracks}
\usage{
\S4method{plot_tracks}{GenomicRanges}(windows, params,
  locus_names = mcols(windows)$name, offset = width(windows[1])\%/\%2,
  xtitle = if (length(windows) > 1) "Relative Position" else
  seqnames(windows), ..., summary_args = list())

\S4method{plot_tracks}{character}(windows, params, locus_names = windows,
  offset = width(windows)\%/\%2, xtitle = if (length(windows) > 1)
  "Relative Position" else seqnames(windows), ..., summary_args = list())
}
\arguments{
\item{windows}{GenomicRanges or rownames for summary object}

\item{params}{TrackParameters, from \code{\link{set_track_parameters}}}

\item{locus_names}{names for each genomic locus represented by windows}

\item{offset}{offset to use for center of region, used when plotting multiple 
regions}

\item{xtitle}{title for x axis}

\item{...}{additional arguments from \code{\link{set_track_parameters}}}

\item{summary_args}{lof arguments to override  from 
\code{\link{set_summary_parameters}}}
}
\value{
GenomeTrackWidgets object, which is displayed as htmlwidgets. To 
convert manually to htmlwidgets, use to_widget
}
\description{
plot_tracks
}
\examples{

library(GenomicRanges)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)

## we'll read in some RNA counts
data(rpkm_chr21)

## From the ranges of the rpkm object, we'll pull out the tss
chr21_promoters <- promoters(SummarizedExperiment::rowRanges(rpkm_chr21),
                 up = 1000, down = 1000)

## set summary parameters

summary_params <- set_summary_parameters(rpkm_chr21,
  groups = "GROUP", ranges = chr21_promoters) 
  
## We'll also read in some track data to plot
genomation_dir <- system.file("extdata", package = "genomationData")
samp.file <- file.path(genomation_dir,'SamplesInfo.txt')
samp.info <- read.table(samp.file, header=TRUE, sep="\\t", 
                        stringsAsFactors = FALSE)
samp.info$fileName <- file.path(genomation_dir, samp.info$fileName)

## Make track plotter using summary parametrs

track_params <- set_track_parameters(samp.info$fileName[1:3], 
  annotation = TxDb.Hsapiens.UCSC.hg19.knownGene, 
  track_names = samp.info$sampleName[1:3], 
  share_y = TRUE,
  summary = summary_params)
  
if (interactive()){
  plot_tracks(rownames(rpkm_chr21)[1:3], track_params)
  plot(chr21_promoters[1:3], track_params)
}   

}
\author{
Alicia Schep and Justin Finkle
}
