% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\docType{methods}
\name{unpack_transcripts}
\alias{unpack_transcripts}
\alias{unpack_transcripts,TxDb-method}
\alias{unpack_transcripts,OrganismDbi-method}
\alias{unpack_transcripts,NULL-method}
\alias{unpack_transcripts,TranscriptParts-method}
\alias{unpack_transcripts,OrganismDb-method}
\title{unpack_transcripts}
\usage{
\S4method{unpack_transcripts}{TxDb}(object)

\S4method{unpack_transcripts}{OrganismDb}(object)
}
\arguments{
\item{object}{a TxDb or OrganismDb object}
}
\value{
TranscriptParts object
}
\description{
Method to load transcript annotation data to a format that is easy to make
transcript visualizations from. The resulting object can be passed to 
make_track_plotter and will result in faster track function creation than 
passing the TxDb or OrganismDb object. Useful if calling make_track_plotter
multiple times.
}
\examples{

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
transc <- unpack_transcripts(TxDb.Hsapiens.UCSC.hg19.knownGene)
class(transc)
}
\author{
Alicia Schep and Justin Finkle
}
