% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\docType{methods}
\name{coverage_heatmap}
\alias{coverage_heatmap}
\alias{coverage_heatmap,matrix-method}
\alias{coverage_heatmap,list-method}
\alias{coverage_heatmap,SummarizedExperiment-method}
\alias{coverage_heatmap,ScoreMatrix-method}
\alias{coverage_heatmap,ScoreMatrixList-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,matrix-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,list-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,SummarizedExperiment-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,ScoreMatrix-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,ScoreMatrixList-method}
\alias{add_coverage_heatmap}
\alias{coverage_heatmap,SummarizedExperiment-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,SummarizedExperiment-method}
\alias{coverage_heatmap,ScoreMatrix-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,ScoreMatrixList-method}
\alias{coverage_heatmap,matrix-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,matrix-method}
\alias{coverage_heatmap,list-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,list-method}
\title{coverage_heatmap}
\usage{
coverage_heatmap(data, ...)

add_coverage_heatmap(p, data, ...)

\S4method{coverage_heatmap}{SummarizedExperiment}(data,
  assay = assayNames(data), ...)

\S4method{add_coverage_heatmap}{IheatmapHorizontal,SummarizedExperiment}(p,
  data, assay = assayNames(data), ...)

\S4method{coverage_heatmap}{ScoreMatrix}(data, start, end, x = seq(start, end,
  length.out = ncol(data)), y = default_y(data), ...)

\S4method{add_coverage_heatmap}{IheatmapHorizontal,ScoreMatrixList}(p, data,
  start, end, x = seq(start, end, length.out = ncol(data)), ...)

\S4method{coverage_heatmap}{matrix}(data, x = default_x(data),
  y = default_y(data), row_order = c("signal", "hclust", "kmeans", "groups",
  "none"), k = NULL, groups = NULL, clust_dist = stats::dist,
  signal = log10rowMeans(data), plot_signal = TRUE, name = "Coverage",
  signal_name = "Avg. (log10)", summary = TRUE,
  scale_method = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), pct = 0.95, scale_factor = 1,
  show_xlabels = TRUE, start = x[1], end = default_end(x),
  col_title = "Position", layout = list(font = list(size = 10)), ...)

\S4method{add_coverage_heatmap}{IheatmapHorizontal,matrix}(p, data,
  x = default_x(data), signal = log10rowMeans(data), plot_signal = TRUE,
  name = "Coverage", signal_name = "Avg. (log10)", summary = TRUE,
  scale_method = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), pct = 0.95, scale_factor = 1,
  show_xlabels = TRUE, start = x[1], end = default_end(x),
  col_title = "Position", ...)

\S4method{coverage_heatmap}{list}(data, x = default_x(data[[1]]),
  y = default_y(data[[1]]), row_order = c("signal", "hclust", "kmeans",
  "groups", "none"), k = NULL, groups = NULL, clust_dist = stats::dist,
  cluster_by = c("first", "all"), signal = lapply(data, log10rowMeans),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. (log10)",
  summary = TRUE, scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, show_xlabels = TRUE, start = x[1],
  end = default_end(x), col_title = "Position", layout = list(font =
  list(size = 10)), ...)

\S4method{add_coverage_heatmap}{IheatmapHorizontal,list}(p, data,
  x = default_x(data[[1]]), signal = lapply(data, log10rowMeans),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. (log10)",
  summary = TRUE, scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, show_xlabels = TRUE, start = x[1],
  end = default_end(x), col_title = "Position", ...)
}
\arguments{
\item{data}{single coverage matrix or list of coverage matrices}

\item{...}{additional arguments}

\item{p}{IHeatmap object}

\item{assay}{name(s) of assay to plot, if data is SummarizedExperiment}

\item{start}{label for start of x range}

\item{end}{label for end of x range}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{clust_dist}{distance function to use for clustering}

\item{signal}{signal along row}

\item{plot_signal}{add an annotation heatmap showing the average signal? 
default is TRUE}

\item{name}{name of colorbar}

\item{signal_name}{name for signal colorbar}

\item{summary}{make summary plot, boolean, default is TRUE}

\item{scale_method}{how to scale matrix before displaying in heatmap}

\item{pct}{percentile to use if scale_method is "PercentileMax"}

\item{scale_factor}{scale_factor to use if scale_method is "scalar"}

\item{show_xlabels}{show xlabels?  default is TRUE}

\item{col_title}{x axis label}

\item{layout}{list of layout attributes}

\item{cluster_by}{use "first" or "all" matrices for clustering if given 
multiple matrices}
}
\value{
iheatmap object
}
\description{
Makes an interactive coverage heatmap.
}
\examples{

library(GenomicRanges)
## First we'll read in some sample data
genomation_dir <- system.file("extdata", package = "genomationData")
samp.file <- file.path(genomation_dir,'SamplesInfo.txt')
samp.info <- read.table(samp.file, header=TRUE, sep='\\t', 
                        stringsAsFactors = FALSE)
samp.info$fileName <- file.path(genomation_dir, samp.info$fileName)
ctcf.peaks = genomation::readBroadPeak(system.file("extdata",
                          "wgEncodeBroadHistoneH1hescCtcfStdPk.broadPeak.gz",
                           package = "genomationData"))
ctcf.peaks = ctcf.peaks[seqnames(ctcf.peaks) == "chr21"]
ctcf.peaks = ctcf.peaks[order(-ctcf.peaks$signalValue)]
ctcf.peaks = resize(ctcf.peaks, width = 501, fix = "center")

## Make coverage matrix
ctcf_mats <- make_coverage_matrix(samp.info$fileName[1:5], 
                                  ctcf.peaks, 
                                  input_names = samp.info$sampleName[1:5],
                                  up = 250, 
                                  down = 250, 
                                  binsize = 25)
                                  
## Plot coverage for Ctcf and Znf143
if (interactive()){
  coverage_heatmap(ctcf_mats, "Ctcf") \%>\% 
    add_coverage_heatmap(ctcf_mats, "Znf143")
}                                   


}
\author{
Alicia Schep
}
