% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{coverage_heatmap}
\alias{coverage_heatmap}
\alias{coverage_heatmap,matrix-method}
\alias{coverage_heatmap,list-method}
\alias{coverage_heatmap,SummarizedExperiment-method}
\alias{coverage_heatmap,ScoreMatrix-method}
\alias{coverage_heatmap,ScoreMatrixList-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,matrix-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,list-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,SummarizedExperiment-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,ScoreMatrix-method}
\alias{add_coverage_heatmap,IheatmapHorizontal,ScoreMatrixList-method}
\alias{add_coverage_heatmap}
\title{coverage_heatmap}
\usage{
coverage_heatmap(data, ...)

add_coverage_heatmap(p, data, ...)
}
\arguments{
\item{data}{single coverage matrix or list of coverage matrices}

\item{...}{additional arguments}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{clust_dist}{distance function to use for clustering}

\item{signal}{signal along row}

\item{plot_signal}{add an annotation heatmap showing the average signal? 
default is TRUE}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean, default is TRUE}

\item{scale_method}{how to scale matrix before displaying in heatmap}

\item{pct}{percentile to use if scale_method is "PercentileMax"}

\item{scale_factor}{scale_factor to use if scale_method is "scalar"}

\item{show_xlabels}{show xlabels?  default is TRUE}

\item{start}{label for start of x range}

\item{end}{label for end of x range}

\item{xlab}{x axis label}

\item{layout}{list of layout attributes}
}
\value{
iheatmap object
}
\description{
Makes an interactive coverage heatmap.
}
\examples{

library(GenomicRanges)
## First we'll read in some sample data
genomation_dir <- system.file("extdata", package = "genomationData")
samp.file <- file.path(genomation_dir,'SamplesInfo.txt')
samp.info <- read.table(samp.file, header=TRUE, sep='\\t', 
                        stringsAsFactors = FALSE)
samp.info$fileName <- file.path(genomation_dir, samp.info$fileName)
ctcf.peaks = genomation::readBroadPeak(system.file("extdata",
                          "wgEncodeBroadHistoneH1hescCtcfStdPk.broadPeak.gz",
                           package = "genomationData"))
ctcf.peaks = ctcf.peaks[seqnames(ctcf.peaks) == "chr21"]
ctcf.peaks = ctcf.peaks[order(-ctcf.peaks$signalValue)]
ctcf.peaks = resize(ctcf.peaks, width = 501, fix = "center")

## Make coverage matrix
ctcf_mats <- make_coverage_matrix(samp.info$fileName[1:5], 
                                  ctcf.peaks, 
                                  input_names = samp.info$sampleName[1:5],
                                  up = 250, 
                                  down = 250, 
                                  binsize = 25)
                                  
## Plot coverage for Ctcf and Znf143
if (interactive()){
  coverage_heatmap(ctcf_mats, "Ctcf") \%>\% 
    add_coverage_heatmap(ctcf_mats, "Znf143")
}                                   


}
\author{
Alicia Schep
}
