% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plus_tracks_shiny.R
\name{heatmap_click}
\alias{heatmap_click}
\title{heatmap_click}
\usage{
heatmap_click(heatmap, x)
}
\arguments{
\item{heatmap}{heatmap object}

\item{x}{list-like object corresponding to rows of heatmap}
}
\value{
function that returns appropriate element of x based on row of 
heatmap clicked within \code{\link{shinyApp}}

returns function
}
\description{
Function to bind heatmap to click event
}
\author{
Alicia Schep
}
