% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plus_tracks_shiny.R
\name{heatmap_to_tracks_shiny}
\alias{heatmap_to_tracks_shiny}
\title{heatmap_to_tracks_shiny}
\usage{
heatmap_to_tracks_shiny(heatmap, track_params, link,
  title = "Heatmap linked to Genome Tracks", options = list(height = 1400))
}
\arguments{
\item{heatmap}{IHeatmap object}

\item{track_params}{output from \code{\link{set_track_parameters}}}

\item{link}{link function, linking rows of heatmap to input to track function
generator, result from \code{\link{heatmap_click}}}

\item{title}{Title of shiny app}

\item{options}{to pass to shiny}
}
\value{
Shiny application
}
\description{
Function for making shiny app linking a heatmap to genome tracks
}
\examples{

library(GenomicRanges)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)

  
## We'll also read in some track data to plot
genomation_dir <- system.file("extdata", package = "genomationData")
samp.file <- file.path(genomation_dir,'SamplesInfo.txt')
samp.info <- read.table(samp.file, header=TRUE, sep="\\t", 
                        stringsAsFactors = FALSE)
samp.info$fileName <- file.path(genomation_dir, samp.info$fileName)
ctcf.peaks = genomation::readBroadPeak(system.file("extdata",
                         "wgEncodeBroadHistoneH1hescCtcfStdPk.broadPeak.gz",
                          package = "genomationData"))
ctcf.peaks = ctcf.peaks[seqnames(ctcf.peaks) == "chr21"]

## resize peaks to size 1000
ctcf.peaks = resize(ctcf.peaks, width = 10000, fix = "center")

## Make track plotter using summary parametrs

track_params <- set_track_parameters(samp.info$fileName[1:3], 
  annotation = TxDb.Hsapiens.UCSC.hg19.knownGene, 
  track_names = samp.info$sampleName[1:3], 
  share_y = TRUE)

# Make coverage heamap

ctcf_mats <- make_coverage_matrix(samp.info$fileName[1:5], 
                                  ctcf.peaks, 
                                  input_names = samp.info$sampleName[1:5],
                                  up = 250, 
                                  down = 250, 
                                  binsize = 25)
                                  
hm <- coverage_heatmap(ctcf_mats, "Ctcf") 

link_fn <- heatmap_click(hm, ctcf.peaks)
      
if (interactive()){
  heatmap_to_tracks_shiny(hm, track_params, link_fn)
}   

}
\author{
Alicia Schep and Justin Finkle
}
