% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_iheatmap.R
\docType{methods}
\name{iheatmap-SummarizedExperiment}
\alias{iheatmap-SummarizedExperiment}
\alias{iheatmap,SummarizedExperiment-method}
\alias{add_iheatmap,IHeatmapHorizontal,SummarizedExperiment-method}
\alias{add_iheatmap,IHeatmapVertical,SummarizedExperiment-method}
\alias{add_iheatmap,IheatmapHorizontal,SummarizedExperiment-method}
\alias{add_iheatmap,IheatmapVertical,SummarizedExperiment-method}
\title{Summarized Experiment iheatmap and iheatmap methods}
\usage{
\S4method{iheatmap}{SummarizedExperiment}(data, assay = assayNames(data)[[1]],
  x = se_default_x(data, assay), y = se_default_y(data, assay),
  cluster_rows = c("hclust", "kmeans", "none"), cluster_cols = c("hclust",
  "kmeans", "none"), scale = "rows", ...)

\S4method{add_iheatmap}{IheatmapHorizontal,SummarizedExperiment}(p, data,
  assay = assayNames(data)[[1]], x = se_default_x(data),
  cluster_cols = c("hclust", "kmeans", "none"), scale = "rows", ...)

\S4method{add_iheatmap}{IheatmapVertical,SummarizedExperiment}(p, data,
  assay = assayNames(data)[[1]], y = se_default_y(data),
  cluster_rows = c("hclust", "kmeans", "none"), scale = "rows", ...)
}
\arguments{
\item{data}{SummarizedExperiment}

\item{assay}{name of assay in data to plot}

\item{x}{name of rows of data, by default rownames}

\item{y}{name of columns of data, by default colnames}

\item{cluster_rows}{how to cluster rows, default is hclust}

\item{cluster_cols}{how to cluster columns, default is hclust}

\item{scale}{scale rows, columns, both, or none}

\item{...}{additional arguments to \code{\link[iheatmapr]{iheatmap}}}

\item{p}{IHeatmap object}
}
\value{
\code{\link[iheatmapr]{Iheatmap-class}} object
}
\description{
iheatmap methods that take as input a SummarizedExperiment object.  With
SummarizedExperiment input, the defaults are different than with a plain 
matrix input, and more likely to be suitable for functional genomics data.
}
\examples{

library(SummarizedExperiment)
data(rpkm_chr21)

hm <- iheatmap(rpkm_chr21, "rpkm",
               x = colData(rpkm_chr21)$STD_NAME, 
               y = rowData(rpkm_chr21)$SYMBOL, 
               col_annotation = colData(rpkm_chr21)[,c("TYPE","SEX")])
               
if (interactive()) {
  hm
}                
}
