% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_matrix.R
\name{make_coverage_matrix}
\alias{make_coverage_matrix}
\title{make_coverage_matrix}
\usage{
make_coverage_matrix(inputs, ranges, input_names = names(inputs),
  binsize = 1, format = c("auto", "bigwig", "bam"), up = 0, down = 0,
  ...)
}
\arguments{
\item{inputs}{filenames of bigwig or bam}

\item{ranges}{ranges for which to compute coverage within}

\item{binsize}{binsize to bin coverage}

\item{format}{format of files, default is auto}

\item{up}{basepairs upstream of center to use}

\item{down}{basepairs downstream of center to use}
}
\value{
RangedSummarizedExperiment
}
\description{
make_coverage_matrix
}
\details{
up and down are 0 by default -- if not specified, actual range is used.  All ranges
must be of equal width.  If up and/or down are provided, then the center of the range and up 
basepairs upstream and down basepairs downstream are used.
}
\author{
Alicia Schep
}
