% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus.R
\name{make_track_plus_summary_plotter}
\alias{make_track_plus_summary_plotter}
\title{Combine genome track view with locus summary}
\usage{
make_track_plus_summary_plotter(track_function, summary_function, windows,
  row_names, summary_width = 0.25)
}
\arguments{
\item{track_function}{function to make tracks, as created by
\code{\link{make_track_plotter}}}

\item{summary_function}{function to make locus summaries, as created by
\code{\link{make_summary_plotter}}}

\item{windows}{GenomicRanges of windows that can serve as inputs to track_function}

\item{row_names}{vector of row names that can serve as inputs to summary_function}

\item{summary_width}{width of summary plot in resulting visualization}
}
\value{
a function which takes in names and generates a plot with track views
and a summary plot per locus
}
\description{
Combine genome track view with locus summary
}
