% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plotCNV}
\alias{plotCNV}
\title{plot SCNA landscape as heatmap}
\usage{
plotCNV(segDf, geneVersion = NULL, sampleOrder = NULL, chrOrder = NULL)
}
\arguments{
\item{segDf}{segment-level SCNA data frame. This is usually the same input as GISTIC2. Make sure there are these columns "sample", "chromosome", "start", "end", "log2".}

\item{geneVersion}{optionally specify the genome version of SCNA data were generated, should be either 'hg19' or 'hg38'. If this option is specified, the function will fill any gaps between the data and the whole genome.
If geneVersion is NULL, the no gap will be filled. Default is NULL.}

\item{sampleOrder}{a optional character vector to specify the sample order for the display. If NULL, the samples will be ranked as their order in the segDf. Default is NULL.}

\item{chrOrder}{a optional character vector to specify the chromosome order for the display. If NULL, the chromosomes will be ranked as their order in the segDf. Default is NULL.}
}
\value{
a ggplot2 plot object
}
\description{
plotCNV is used to plot the segment-level SCNA data into heatmap
}
\examples{
hnsccSegfile <- system.file("extdata", "CPTAC3_HNSCC_SCNA_segment_level.tsv", package = "genomicWidgets")
hnsccSegDf <- read.table(hnsccSegfile, header = TRUE)
plotCNV(hnsccSegDf, geneVersion = 'hg38')
}
