% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus.R
\docType{methods}
\name{to_widget,LocusViewList-method}
\alias{to_widget,LocusViewList-method}
\alias{to_widget,LocusSummaryList-method}
\alias{to_widget}
\alias{to_widget,GenomeTrackWidget-method}
\alias{to_widget,NULL-method}
\alias{to_widget,LocusViewList-method}
\alias{to_widget,LocusView-method}
\alias{to_widget,NULL-method}
\title{to_widget}
\usage{
\S4method{to_widget}{LocusViewList}(p)

\S4method{to_widget}{LocusSummaryList}(p)

\S4method{to_widget}{GenomeTrackWidget}(p)

\S4method{to_widget}{`NULL`}(p)
}
\arguments{
\item{p}{GenomeTrackWidget or other object storing plot information}
}
\value{
htmlwidgets object
}
\description{
Method to convert GenomeTrackWidget to htmlwidgets objects
}
