% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus.R
\docType{methods}
\name{to_widget}
\alias{to_widget}
\title{to_widget}
\usage{
\S4method{to_widget}{GenomeTrackWidget}(p)
}
\arguments{
\item{p}{GenomeTrackWidget or other object storing plot information}
}
\value{
htmlwidgets object
}
\description{
Method to convert GenomeTrackWidget to htmlwidgets objects
}
